+++ libs/core/CHANGELOG.md
@@ -1,5 +1,63 @@
 # Changelog
 
+## 1.88.3
+
+### Patch Changes
+
+- e480d9f: **Checkbox:** Fix interaction issue with NVDA. Fixes #2199
+- 1331668: **Checkbox:** Add missing dependency of `IconMinusSmall`
+
+## 1.88.2
+
+### Patch Changes
+
+- 72be834: **Dropdown:** Fix issue with updating error message via prop
+
+## 1.88.1
+
+### Patch Changes
+
+- 0edf04f: **Calendar:** Add disabled class on weeknumber
+
+## 1.88.0
+
+### Minor Changes
+
+- 27fe9cf: **Core**: Spinner component label position prop
+
+## 1.87.0
+
+### Minor Changes
+
+- f3b8f76: **Icon:** add new code icons
+
+### Patch Changes
+
+- 577fc63: Fix, calculating extended supporting text height on resize
+- eca17f0: **Breadcrumbs:** Fix custom html scoping
+
+## 1.86.0
+
+### Minor Changes
+
+- 1ac544d: **Popover:** Close popover if page starts scrolling. Closes #2134
+
+### Patch Changes
+
+- f247f14: **Focus handling:** Fix issues with programmatic focus shifting with some form controls
+
+## 1.85.4
+
+### Patch Changes
+
+- c7610e3: FAB was missing from the index.ts export, meaning you have to import FAB from a subfolder, causing two import lines from green-core.
+
+## 1.85.3
+
+### Patch Changes
+
+- 1891932: Add missing exports from alert, rich-text and segment.
+
 ## 1.85.2
 
 ### Patch Changes